deployment/aurora-standard-node/services/slack-integration/requirements.txt
slack-bolt==1.18.1
slack-sdk==3.27.1
fastapi==0.104.1
uvicorn==0.24.0
psycopg2-binary==2.9.9
redis==5.0.1
pydantic==2.5.0
httpx==0.25.2
aiohttp==3.9.1
asyncpg==0.29.0
